Algorithm to generate growth of plants using core JAVA OOPS concepts, multi threading, JAVA Graphics
Algorithm to generate growth of plants using core JAVA OOPS concepts, multi threading, JAVA Graphics. BGStem is a brach intial for whole tree. BGGeneration represents set of BGStems which leads to one generations. BGGenerationset contains set of multiple generations which forms tree. It uses BGRule class to use set of rules based on which it generates the generations. Uses JAVA Graphics to display the output on the screen.